backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.61.0
anyio==4.10.0
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.1
azure-ai-agents==1.1.0
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b12
azure-common==1.1.28
azure-core==1.35.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.14.0b1
azure-identity==1.23.1
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
croniter==6.0.0
cryptography==45.0.6
dataclasses-json==0.6.7
debtcollector==3.0.0
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dogpile.cache==1.4.0
eval_type_backport==0.2.2
eventlet==0.40.2
fastapi==0.116.1
fasteners==0.19
filelock==3.18.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
futurist==3.2.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==8.7.0
invoke==2.2.0
iso8601==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keystoneauth1==5.11.1
keystonemiddleware==10.11.0
kombu==5.5.4
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-azure-ai==0.1.5
langchain-community==0.3.27
langchain-core==0.3.72
langchain-mcp-adapters==0.1.9
langchain-mistralai==0.2.11
langchain-ollama==0.3.6
langchain-openai==0.3.28
langchain-postgres==0.0.15
langchain-text-splitters==0.3.9
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langsmith==0.4.13
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.12.4
mistral==20.0.0
mistral-lib==3.4.0
mistralai==1.9.3
msal==1.33.0
msal-extensions==1.3.1
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
netaddr==1.3.0
networkx==3.5
numpy==2.3.2
ollama==0.5.2
openai==1.99.1
opentelemetry-api==1.36.0
orjson==3.11.1
ormsgpack==1.10.0
os-service-types==1.8.0
oslo.cache==3.11.0
oslo.concurrency==7.1.0
oslo.config==10.0.0
oslo.context==6.0.0
oslo.db==17.3.0
oslo.i18n==6.5.1
oslo.log==7.2.0
oslo.messaging==17.1.0
oslo.metrics==0.12.0
oslo.middleware==6.6.0
oslo.policy==4.5.1
oslo.serialization==5.7.0
oslo.service==4.2.2
oslo.utils==9.0.0
osprofiler==4.3.0
packaging==25.0
paramiko==4.0.0
passlib==1.7.4
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.1
pdf2image==1.17.0
pecan==1.7.0
pgvector==0.3.6
pillow==11.3.0
ply==3.11
prettytable==3.16.0
prometheus_client==0.22.1
propcache==0.3.2
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pycadf==4.0.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-keystoneclient==5.6.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
repoze.lru==0.7
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.27.0
simplegeneric==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sse-starlette==3.0.2
starlette==0.47.2
statsd==4.0.1
stevedore==5.4.1
tenacity==9.1.2
testresources==2.0.2
testscenarios==0.5.0
testtools==2.7.2
tiktoken==0.10.0
tokenizers==0.21.4
toml==0.10.2
tooz==7.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
WebOb==1.8.9
Werkzeug==3.1.3
wrapt==1.17.2
WSME==0.12.1
xxhash==3.5.0
yappi==1.6.10
yaql==3.1.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
